feat: getuUserReview 연결
#98 #101

+++ app/src/main/java/com/d_vide/D_VIDE/app/presentation/Followings/FollowViewModel.kt
@@ -58,7 +58,7 @@ class FollowViewModel @Inject constructor(
         }.launchIn(viewModelScope)
     }
 
-    fun postFollow(userId: Int){
+    fun postFollow(userId: Long){
         _userIdDTO.value = userIdDTO.value.copy(
             userId = userId.toLong()
         )
@@ -71,7 +71,7 @@ class FollowViewModel @Inject constructor(
         }.launchIn(viewModelScope)
     }
 
-    fun deleteFollow(userId: Int){
+    fun deleteFollow(userId: Long){
         _userIdDTO.value = userIdDTO.value.copy(
             userId = userId.toLong()
         )
